Refactored project structure to seperate functions
