+++ .gitignore
@@ -351,3 +351,6 @@ MigrationBackup/
 
 # Rider
 .idea/
+
+# Azure
+**/local.settings.json
