+++ src/Sito.QR.Api.Generators/Factories/Abstractions/IPayloadFactory.cs
@@ -1,6 +1,6 @@
-using Sito.QR.Generators.Shared.Dto;
+using Sito.QR.Api.Shared.Dto;
 
-namespace Sito.QR.Api.Factories.Abstractions;
+namespace Sito.QR.Api.Generators.Factories.Abstractions;
 
 public interface IPayloadFactory<in TDto> where TDto : IQrRequest
 {
