+++ src/Sito.QR.Api.Shared/Sito.QR.Api.Shared.csproj
@@ -1,0 +1,19 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net6.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\Sito.QR.Shared\Sito.QR.Shared.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <PackageReference Include="Azure.Messaging.ServiceBus" Version="7.7.0" />
+      <PackageReference Include="Microsoft.Azure.Functions.Extensions" Version="1.1.0" />
+      <PackageReference Include="Microsoft.Extensions.Azure" Version="1.1.1" />
+    </ItemGroup>
+
+</Project>
