+++ src/Sito.QR.Api/Functions/QrApi.cs
@@ -8,7 +8,7 @@ using Microsoft.Azure.WebJobs.Extensions.Http;
 using Microsoft.Extensions.Logging;
 using Sito.QR.Api.Factories.Abstractions;
 using Sito.QR.Api.Repositories.Abstractions;
-using Sito.QR.Generators.Shared.Dto;
+using Sito.QR.Api.Shared.Dto;
 
 namespace Sito.QR.Api.Functions;
