--- src/Sito.QR.Generators.Shared/Dto/IQrRequest.cs
@@ -1,8 +1,0 @@
-using Sito.QR.Generators.Shared.Enum;
-
-namespace Sito.QR.Generators.Shared.Dto;
-
-public interface IQrRequest
-{
-    QrType Type { get; }
-}
