+++ src/Sito.QR.Shared/Dto/IQrRequest.cs
@@ -1,0 +1,8 @@
+using Sito.QR.Api.Shared.Enum;
+
+namespace Sito.QR.Api.Shared.Dto;
+
+public interface IQrRequest
+{
+    QrType Type { get; }
+}
