+++ src/Sito.QR.Shared/Sito.QR.Shared.csproj
@@ -4,6 +4,7 @@
         <TargetFramework>net6.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
+        <RootNamespace>Sito.QR.Api.Shared</RootNamespace>
     </PropertyGroup>
 
 </Project>
