+++ src/Sito.QR.sln
@@ -2,9 +2,15 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sito.QR.Api", "Sito.QR.Api\Sito.QR.Api.csproj", "{B623AF77-027A-43C1-89C0-3D5E7A9A3414}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Generators", "Generators", "{22F17D37-7658-4922-8AD8-CE7E5DA03C83}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Functions", "Functions", "{22F17D37-7658-4922-8AD8-CE7E5DA03C83}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sito.QR.Generators.Shared", "Sito.QR.Generators.Shared\Sito.QR.Generators.Shared.csproj", "{8D202FF8-4DDC-4D3A-BEEF-E7365056615B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sito.QR.Shared", "Sito.QR.Shared\Sito.QR.Shared.csproj", "{8D202FF8-4DDC-4D3A-BEEF-E7365056615B}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{BC02E190-0466-4E02-A1AD-A8A3964D6BD2}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sito.QR.Api.Generators", "Sito.QR.Api.Generators\Sito.QR.Api.Generators.csproj", "{60DB3DAA-FF86-46FA-A1C5-B8E61E0FDD5A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sito.QR.Api.Shared", "Sito.QR.Api.Shared\Sito.QR.Api.Shared.csproj", "{D5AE98BF-3D1E-4D72-A73C-377F7A5C1AD6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -20,9 +26,18 @@ Global
 		{8D202FF8-4DDC-4D3A-BEEF-E7365056615B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8D202FF8-4DDC-4D3A-BEEF-E7365056615B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8D202FF8-4DDC-4D3A-BEEF-E7365056615B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{60DB3DAA-FF86-46FA-A1C5-B8E61E0FDD5A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{60DB3DAA-FF86-46FA-A1C5-B8E61E0FDD5A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{60DB3DAA-FF86-46FA-A1C5-B8E61E0FDD5A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{60DB3DAA-FF86-46FA-A1C5-B8E61E0FDD5A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D5AE98BF-3D1E-4D72-A73C-377F7A5C1AD6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D5AE98BF-3D1E-4D72-A73C-377F7A5C1AD6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D5AE98BF-3D1E-4D72-A73C-377F7A5C1AD6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D5AE98BF-3D1E-4D72-A73C-377F7A5C1AD6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{B623AF77-027A-43C1-89C0-3D5E7A9A3414} = {22F17D37-7658-4922-8AD8-CE7E5DA03C83}
-		{8D202FF8-4DDC-4D3A-BEEF-E7365056615B} = {22F17D37-7658-4922-8AD8-CE7E5DA03C83}
+		{60DB3DAA-FF86-46FA-A1C5-B8E61E0FDD5A} = {22F17D37-7658-4922-8AD8-CE7E5DA03C83}
+		{D5AE98BF-3D1E-4D72-A73C-377F7A5C1AD6} = {22F17D37-7658-4922-8AD8-CE7E5DA03C83}
 	EndGlobalSection
 EndGlobal
